+++ cmd/gravinet/main.go
@@ -48,7 +48,7 @@ import (
 
 // Build metadata, overridable via -ldflags.
 var (
-	version = "729"
+	version = "733"
 	commit  = "none"
 )